+++ homeassistant/components/dialogflow/strings.json
@@ -1,0 +1,18 @@
+{
+  "config": {
+    "title": "Dialogflow",
+    "step": {
+      "user": {
+        "title": "Set up the Dialogflow Webhook",
+        "description": "Are you sure you want to set up Dialogflow?"
+      }
+    },
+    "abort": {
+      "one_instance_allowed": "Only a single instance is necessary.",
+      "not_internet_accessible": "Your Home Assistant instance needs to be accessible from the internet to receive Dialogflow messages."
+    },
+    "create_entry": {
+      "default": "To send events to Home Assistant, you will need to setup [webhook integration of Dialogflow]({dialogflow_url}).\n\nFill in the following info:\n\n- URL: `{webhook_url}`\n- Method: POST\n- Content Type: application/json\n\nSee [the documentation]({docs_url}) for further details."
+    }
+  }
+}
