+++ tests/components/dialogflow/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for the Dialogflow component."""
